\name{data0}
\alias{demo}
\alias{data}
\docType{data}
\title{
DEMO
}
\description{
A demo for ddmarker and MMC. \cr
}
\usage{data(data0)}
\format{
  The format is:
 Large list [10 elements] ...
 - attr(*, "dimnames")=List of 10
  ..$demo.pm : "numeric" ...
  ..$demo.ddm : "character" ...
  ..$demo.pro : "character" ...
  ..$demo.gen : "character" ...
  ..$demo.iso : "character" ...
  ..$demo.ent : "character" ...
  ..$demo.mir : "character" ...
  ..$demo.seq : "character" ...
  ..$MI : "numeric" ...
  ..$MP : "numeric" ...
}
\examples{
data(data0);
\code{\link{DDMarker-method}} \code{\link{MMC-method}} 
}
\author{
Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
Qiong Yu (JLU & UGA) \email{yuqiong@uga.edu} \cr 
Huansheng Cao (UGA) \email{hshcao@uga.edu} \cr 
Guoqing Liu (IMUST & UGA) \email{gqliu@uga.edu} \cr 
Wei Du (JLU & UGA) \email{weidu@uga.edu} \cr 
Yan Wang (JLU & UGA) \email{wy6868@hotmail.com} \cr 
Hao Wu (BIT & UGA) \email{wuhao@uga.edu} \cr 
Jingyu Yang (GSUT & UGA) \email{jyyyjy@uga.edu} \cr 
Xiufeng Liu (GZUCM & UGA) \email{xfliu@uga.edu} \cr 
Yuan Guo (SAHCQMU & UGA) \email{yuanguo@uga.edu} \cr 
Ying Xu (JLU & UGA) \email{xyn@bmb.uga.edu} \cr \cr 
Maintainer: Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
}
\references{
citation("DDMarker");
\cr 
\code{[1]} Yan Wang, et al. (2009) \emph{DMarker: A Bio-Marker Inference System for Human Diseases based
on Microarray Gene Expression Data} 2009 \cr
\code{[2]} Juan Cui, et al. (2008) \emph{Computational prediction of human proteins that can be secreted into the bloodstream} BIOINFORMATICS, Vol.24 no. 20 2008 pages 2370-2375 \cr
\code{[3]} Jiaxin Wang, et al. (2013) \emph{Computational Prediction of Human Salivary Proteins from Blood Circulation and Application to Diagnostic Biomarker Identification} PLoS ONE, DOI: 10.1371/journal.pone.0080211, 2013 \cr
\code{[4]} S Hong, et al. (2011) \emph{A Computational Method for Prediction of Excretory Proteins and Application to Identification of cancer markers in urine and application to gastric cancer} PLoS ONE,6(2):e16875, 2011 \cr
\code{[5]} \url{http://bioinfosrv1.bmb.uga.edu/DMarker/} \cr
}
\keyword{ datasets }
\keyword{ demo }
\keyword{ matrix }
\seealso{
\code{\link{DDMarker-package}} \cr
}
