\name{DDMarker-package}
\alias{DDMarker-package}
\docType{package}
\title{
DDMarker
}
\description{
Diagnose and Detect Markers in Extracellular Circulating is a homo sapiens deductive system solving the markers in extracellular circulating. It entails the symbols of markers, like the genes, the proteins, the micro RNAs, and the isoforms, whether can be diagnose and detect in extracellular circulating, especially the blood serum and the urine for the biological and medicine significance. With the help of a homo sapiens annotation database in DDMarkerData package, DDMarker can even diagnose and detect the sequence among the genes, the proteins, the micro RNAs, and the isoforms. There are two main function in this package, \code{the ddmarker}, and \code{the DDMarkerMMC}, short for Minimal Metabolize Circulation. \code{DDMarkerMMC} entails the markers among the minimal metabolize circulation. \cr
The main function of the package are \code{ddmarker()} and \code{DDMarkerMMC()}. The more details you can find in \code{\link{DDMarker-method}} and \code{\link{MMC-method}} \cr
}
\details{
\tabular{ll}{
Package: \tab DDMarker\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-27\cr
Depends: \tab R (>= 3.0.3), Rcpp (>= 0.11.3)\cr
LinkingTo: \tab Rcpp\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
LazyData: \tab true\cr
}
}
\author{
Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
Qiong Yu (JLU & UGA) \email{yuqiong@uga.edu} \cr 
Huansheng Cao (UGA) \email{hshcao@uga.edu} \cr 
Guoqing Liu (IMUST & UGA) \email{gqliu@uga.edu} \cr 
Wei Du (JLU & UGA) \email{weidu@uga.edu} \cr 
Yan Wang (JLU & UGA) \email{wy6868@hotmail.com} \cr 
Hao Wu (BIT & UGA) \email{wuhao@uga.edu} \cr 
Jingyu Yang (GSUT & UGA) \email{jyyyjy@uga.edu} \cr 
Xiufeng Liu (GZUCM & UGA) \email{xfliu@uga.edu} \cr 
Yuan Guo (SAHCQMU & UGA) \email{yuanguo@uga.edu} \cr 
Ying Xu (JLU & UGA) \email{xyn@bmb.uga.edu} \cr \cr 
Maintainer: Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
}
\references{
citation("DDMarker");
\cr 
\code{[1]} Yan Wang, et al. (2009) \emph{DMarker: A Bio-Marker Inference System for Human Diseases based
on Microarray Gene Expression Data} 2009 \cr
\code{[2]} Juan Cui, et al. (2008) \emph{Computational prediction of human proteins that can be secreted into the bloodstream} BIOINFORMATICS, Vol.24 no. 20 2008 pages 2370-2375 \cr
\code{[3]} Jiaxin Wang, et al. (2013) \emph{Computational Prediction of Human Salivary Proteins from Blood Circulation and Application to Diagnostic Biomarker Identification} PLoS ONE, DOI: 10.1371/journal.pone.0080211, 2013 \cr
\code{[4]} S Hong, et al. (2011) \emph{A Computational Method for Prediction of Excretory Proteins and Application to Identification of cancer markers in urine and application to gastric cancer} PLoS ONE,6(2):e16875, 2011 \cr
\code{[5]} \url{http://bioinfosrv1.bmb.uga.edu/DMarker/} \cr
}
\keyword{ DDMarker }
\keyword{ MMC }
\keyword{ EC }
\seealso{
\code{\link{DDMarker-method}} \code{\link{MMC-method}} \code{\link{EC}} \code{\link{demo}} \cr
}
